Group generic sharing UI under components/share/
Sharing had grown into several components scattered across card-pieces/,
playlists/, and the top level — two families (send-in-chat and native/OS
share) with no shared home, so the cluster wasn't legible as one capability.
Following the codebase's area-folder convention, move the generic,
cross-app sharing components into components/share/ and name files by role
(the folder is the namespace):

  friend-picker            (was select-friends-to-share)
  send-phrase/request/playlist   (was card-pieces|playlists/send-*-to-friend)
  native-share-button, share-phrase/request/playlist-button

Kept both verbs deliberately: "send" = in-app chat, "share" = native OS
sheet — a meaningful distinction, not folder-redundancy. No barrel (matches
every other component folder). Data stays in features/social
(useSendToFriends); this folder is presentation only.

ReviewShareButton stays in components/review/ — it isn't a generic wrapper
but review-domain logic (the Wordle-style emoji grid from useReviewsToday)
that merely ends in a share, so it's colocated with its data.

Also renamed the headline export SelectFriendsToShareDialog -> FriendPickerDialog
to match its file. Pure move + import-path/string updates; no behavior change.
