+++ src/components/playlists/playlist-list-item.tsx
@@ -5,8 +5,8 @@ import { PhrasePlaylistType } from '@/features/playlists/schemas'
 import { UidPermalink } from '@/components/card-pieces/user-permalink'
 import { Badge, LangBadge } from '@/components/ui/badge'
 import { useOnePlaylistPhrases } from '@/features/playlists/hooks'
-import { SharePlaylistButton } from './share-playlist-button'
-import { SendPlaylistToFriendDialog } from './send-playlist-to-friend'
+import { SharePlaylistButton } from '@/components/share/share-playlist-button'
+import { SendPlaylistToFriendDialog } from '@/components/share/send-playlist'
 import { UpvotePlaylist } from './upvote-playlist-button'
 import { Link } from '@tanstack/react-router'
 import { PlaylistEmbed, isEmbeddableUrl } from './playlist-embed'
