+++ src/routes/_user/friends/invite.lazy.tsx
@@ -11,7 +11,7 @@ import { QRCodeSVG } from 'qrcode.react'
 
 import { Card, CardContent, CardHeader } from '@/components/ui/card'
 import { useProfile } from '@/features/profile/hooks'
-import { NativeShareButton } from '@/components/native-share-button'
+import { NativeShareButton } from '@/components/share/native-share-button'
 import CopyLinkButton from '@/components/copy-link-button'
 import { copyLink } from '@/lib/utils'
