+++ src/routes/themes.index.lazy.tsx
@@ -55,9 +55,9 @@ import { statusStrings } from '@/components/card-pieces/card-status-dropdown'
 import { Avatar, AvatarFallback } from '@/components/ui/avatar'
 import { Button } from '@/components/ui/button'
 import {
-	SelectFriendsToShareDialog,
+	FriendPickerDialog,
 	SharePreviewChip,
-} from '@/components/select-friends-to-share'
+} from '@/components/share/friend-picker'
 import type { RelationsFullType } from '@/features/social/live'
 import Callout from '@/components/ui/callout'
 import { ChoiceTile } from '@/components/ui/choice-tile'
@@ -1013,7 +1013,7 @@ function ShowcaseFriendPicker() {
 			<Button onClick={() => setOpen(true)} className="gap-2">
 				<Send className="size-4" /> Send in chat
 			</Button>
-			<SelectFriendsToShareDialog
+			<FriendPickerDialog
 				open={open}
 				onOpenChange={setOpen}
 				friends={SHOWCASE_FRIENDS}
